make on_message_create event better :)

+++ main.py
@@ -122,17 +122,13 @@ async def on_ready():
 
 @bot.event
 async def on_message(message: discord.Message):
-    if bot.user.mentioned_in(message):
-        await message.add_reaction("👋")
+    if bot.user.mentioned_in(message): await message.add_reaction("👋")
     
-    if message.content == "(please do consider using vxtwitter please and thank you)" and message.author.id == 1091826653367386254:
-        await message.delete()
+    if message.content == "(please do consider using vxtwitter please and thank you)" and message.author.id == 1091826653367386254: await message.delete()
 
-    if ":pepeAngryPing:"in  message.content and message.author.id == 1091826653367386254:
-        await message.delete()
+    if ":pepeAngryPing:"in  message.content and message.author.id == 1091826653367386254: await message.delete()
 
-    if "Sent to https://bloopertrack.club" in message.content and message.author.id == 1091826653367386254:
-        await message.delete()
+    if "Sent to https://bloopertrack.club" in message.content and message.author.id == 1091826653367386254: await message.delete()
 
     await bot.process_commands(message)
